Added a cache variable to force build submodule projects
Defaults to "build", ensures that the build and install steps are run
for the main projects. If a configure is necessary then this will also
be triggered from the build of the external project.

With Visual Studio, default to running "configure" of each submodule
with every super build, since it does not detect changes to submodule
CMakeLists files.

"install" is also a valid choice, and "OFF" turns off
the force build of submodules altogether.

Change-Id: Ie8f70406bb3a6c30f5acb23a73eea22402459c61