Update boost and mongo to more recent versions
mongo is a tmp snapshot to get a bug fix related to updating to use
boost filesystem v3 instead of v2. We'll update to an official release
as soon as there is one that incorporates this bug fix.

Boost is updated so that they both still work together even when using
gcc 4.7 or later.

Change-Id: I23fa01122811101b4c60b81f9fb4d4e171d32b79