Turn off native CPU compiler optimizations, versioned plugins
The native CPU compiler optimizations lead to inconsistent builds that
are optimized for the CPU on the build machine. This can lead to runtime
errors that are very difficult to debug/diagnose. It took quite a while
to get to the bottom of packaging issues on the Mac due to this flag...

Turn off versioned plugins to simplify our runtime code, we only need a
single version in our packages, and this generally just makes our runtime
logic more susceptible to errors.