Added USE_VTK CMake option to root Open Chemistry
Newer updates to avogadrolibs and avogadroapp have a USE_VTK option
that is set to OFF by default. This update allows the user to
specify whether to use VTK or not from the root Open Chemistry
directory (before: the user would have to change the options in
the CMakeLists.txt files in avogadrolibs and avogadroapp manually).
In addition, if USE_VTK is OFF and Mongo Chem is not being built,
VTK is not compiled. This speeds up compilation-from-scratch time
significantly when VTK is not being used (and there does not seem
to be a reason to compile it if it is not being used).