Fixed msym.dll install location on Windows
For some reason (unknown to me), when msym is being compiled,
${INSTALL_BIN_DIR} is set to be ${CMAKE_INSTALL_PREFIX}
instead of ${CMAKE_INSTALL_PREFIX}/bin. This results in the
msym.dll file benig incorrectly installed in ${CMAKE_INSTALL_PREFIX}.
This then results in fixup_bundle not being able to find msym.dll.

It is easy to do a quick-fix: you just have to copy msym.dll into
bin or lib, and fixup_bundle will find it. However, it is somewhat of
a nuisance. This PR fixes the issue. Passing this cmake definition to
msym causes ${INSTALL_BIN_DIR} to be ${CMAKE_INSTALL_PREFIX}/bin, so
msym.dll is installed there and fixup_bundle has no troubles.