Update to Open Babel 3.0.0
The MAE parser adds a Boost dependency, and JSON adds rapid JSON, so
disabling both for now. This seems to build as expected, and I was able
to load files using Open Babel and optimize the geometry.

One point of concern is that the formats seems to be in lib/openbabel
now, and the data in share/openbabel/3.0.0 which has the version despite
me using the flag asking to disable it. This may need some updates to
Avogadro to find the appropriate directories and pass them to obabel in
our relocatable package.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>