Fix cmake warnings
Support for using a cmake versions < 3.0 is announced to be dropped in
the not too distant future and since the sub-projects use more recent
cmake versions anyway, lifting the minimum required version to 3.0 for
the super-project shouldn't hurt (and it removes a warning).

When using cmake 3.24 or more recent there will be a warning about
policy CMP0135 not being set. Thus, we simply check if the current cmake
version knows this policy and if it does, we set it. That way the
warning is silenced.

Signed-off-by: Robert Adam <dev@robert-adam.de>