Update the Open Babel SHA to fix a build failure
Compiling on Arch with all the latest updates was failing for me,
looking into it there is already a PR merged that resolves the issue.
Moving Open Babel to the current master to pull that change in as there
are no recent releases. We may want to tweak the SHA if it causes other
issues.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>